Fix concurrent NSMutableSet modification bug by using a single managed object context
Remove the backgroundContext which was being used at the same time as the
primary context and use the primary context everywhere. It seems the
background context was a vesitge of the pre NSPrivateQueueConcurrencyType
era. -recordsFromMMRecordResponse was already doing a performBlockAndWait
on its given context so we don't need multiple contexts to be safe.

Remove an earlier performBlockAndWait in MMRecordResponse we had added to
to fix other bugs caused by using two different contexts. The single
performBlockAndWait in AFMMRecordResponseSerializer now takes care of this
because we are using the same context.